Shoping_site
I developed a simple e-commerce site using Python and Django as the main technologies for the back-end, along with HTML and Tailwind CSS for the front-end. Python and Django provide a robust and flexible framework to handle the site's functionality, while HTML and Tailwind CSS allow for an elegant and responsive user interface.